Moving some header files which have been copied from other repos/projects into the 'borrowed' directory
