+++ cpp/src/parquet/encryption/external/borrowed/dbpa_interface.h
