+++ cpp/src/parquet/encryption/external/borrowed/enums.h
