+++ cpp/src/parquet/encryption/external/borrowed/span.hpp
