+++ cpp/src/parquet/encryption/external/dbpa_utils.h
@@ -2,7 +2,7 @@
 
 #include <stdexcept>
 
-#include "parquet/encryption/external/dbpa_interface.h"
+#include "parquet/encryption/external/borrowed/dbpa_interface.h"
 #include "parquet/types.h"
 #include "arrow/type_fwd.h" // For arrow::Compression
